update
데미지 스케일이 적용이 안됬던 문제 수정
캐릭터가 어느 상태이든 Hp가 없으면 바로 DEAD 애니메이션이 나오도록 변경
말단 해적 포격의 포물선 높이와 포탄속도는 거리에 비례해서 증가/감소

+++ Assets/Scripts/02_Base/Projectile.cs
@@ -18,7 +18,7 @@ public class Projectile : BaseObject, IPoolable
         StartCoroutine(ShootStraightCoroutine(direction, moveSpeed, lifeTime));
     }
 
-    public void ShootParabola(Vector3 startPosition, Vector3 endPosition, float height, int lifeTime)
+    public void ShootParabola(Vector3 startPosition, Vector3 endPosition, float height, float lifeTime)
     {
         StartCoroutine(ShootParabolaCoroutine(startPosition, endPosition, height, lifeTime));
     }
@@ -96,7 +96,7 @@ public class Projectile : BaseObject, IPoolable
         StageManager.Instance.PoolSystem.Release(this);
     }
 
-    IEnumerator ShootParabolaCoroutine(Vector3 startPosition, Vector3 endPosition, float height, int lifeTime)
+    IEnumerator ShootParabolaCoroutine(Vector3 startPosition, Vector3 endPosition, float height, float lifeTime)
     {
         float timer = 0f;
         m_RigidBody.isKinematic = false;
